// src/types/YouTubeSearchType.ts
export interface YouTubeSearchType {
    type: YouTubeSearchResultType;
    author: string;
    authorId: string;
    authorUrl: string;
    authorVerified: boolean;
    authorThumbnails?: YouTubeSearchThumbnailType[];
    autoGenerated?: boolean;
    subCount?: number;
    videoCount?: number;
    channelHandle?: string;
    description: string;
    descriptionHtml: string;
    title?: string;
    videoId?: string;
    videoThumbnails?: YouTubeSearchThumbnailType[];
    viewCount?: number;
    viewCountText?: string;
    published?: number;
    publishedText?: string;
    lengthSeconds?: number;
    liveNow?: boolean;
    premium?: boolean;
    isUpcoming?: boolean;
}

export interface YouTubeSearchThumbnailType {
    url: string;
    width: number;
    height: number;
    quality?: Quality;
}

export enum Quality {
    Default = "default",
    End = "end",
    High = "high",
    Maxres = "maxres",
    Maxresdefault = "maxresdefault",
    Medium = "medium",
    Middle = "middle",
    Sddefault = "sddefault",
    Start = "start",
}

export enum YouTubeSearchResultType {
    Channel = "channel",
    Video = "video",
}